requirements.txt
Django
djangorestframework
psycopg2
drf-spectacular
Pillow
uwsgi